move to legacy: weavejs_fixes
