+++ weave/legacy/op_def.py
@@ -31,7 +31,7 @@ from weave.legacy.language_features.tagging import (
     tagged_value_type,
 )
 from weave.legacy.run import Run
-from weave.weavejs_fixes import fixup_node
+from weave.legacy.weavejs_fixes import fixup_node
 
 if typing.TYPE_CHECKING:
     from weave.trace import weave_client
