+++ weave/tests/legacy/test_storage.py
@@ -13,8 +13,8 @@ from weave.legacy.arrow import list_ as arrow
 from ... import api as weave
 from ... import storage
 from ... import weave_types as types
+from ...legacy.weavejs_fixes import recursively_unwrap_unions
 from ...weave_internal import make_const_node
-from ...weavejs_fixes import recursively_unwrap_unions
 from . import test_helpers
