+++ weave/weave_server.py
@@ -35,9 +35,8 @@ from weave import (
     server,
     storage,
     util,
-    weavejs_fixes,
 )
-from weave.legacy import context_state, graph, value_or_error, wandb_api
+from weave.legacy import context_state, graph, value_or_error, wandb_api, weavejs_fixes
 from weave.legacy.language_features.tagging import tag_store
 from weave.server_error_handling import client_safe_http_exceptions_as_werkzeug
